requirements.txt
scrapy==2.10.1
scrapy_user_agents==0.1.1
pandas==2.1.0
pandas-gbq==0.19.2
google-cloud-bigquery==3.11.4
google-auth==2.22.0
cloudscraper==1.2.71
lxml==4.9.3
pypiwin32==223
tkhtmlview==0.2.0

cfscrape==2.1.1
nltk==3.8.1